Unified walk speeds across platform: Implemented the walk to use repeatedly execute, so the walking speed will be the same across platforms (depends on FPS).

+++ API/Objects/Characters/Walking/IWalkBehavior.cs
@@ -4,7 +4,7 @@ namespace AGS.API
 {
 	public interface IWalkBehavior : IComponent
 	{
-		int WalkSpeed { get; set; }
+		float WalkSpeed { get; set; }
         bool AdjustWalkSpeedToScaleArea { get; set; }
 		bool IsWalking { get; }
